move property vistor codes to device_instance_types
